chore: made new files for Employee and EmployeeManager to separate tests
